requirements.txt
boto3==1.26.106
psycopg2-binary==2.9.9
